+++ .gitignore
@@ -1,3 +1,9 @@
 target
 Cargo.lock
 
+
+
+# Added by cargo
+
+/target
+/Cargo.lock
