--- crates/pubnub/Cargo.toml
@@ -1,22 +1,0 @@
-[package]
-name = "pubnub"
-version = "0.0.0"
-edition = "2021"
-license = "MIT"
-authors = ["PubNub <support@pubnub.com>"]
-description = "PubNub SDK for Rust"
-repository = "https://github.com/pubnub/rust"
-documentation = "https://docs.rs/pubnub" # TODO: add docs and connect with pubnub docs
-homepage = "https://www.pubnub.com"
-categories = ["api-bindings", "asynchronous", "network-programming"]
-
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-derive_builder = "0.12.0"
-pubnub_core = { path = "../pubnub_core" }
-
-[dev-dependencies]
-async-trait = "0.1"
-tokio = { version = "1", features = ["rt", "macros"]}
