--- crates/pubnub/README.md
@@ -1,14 +1,0 @@
-# Pubnub
-
-Main crate of the Pubnub SDK.
-
-## Support
-
-If you **need help** or have a **general question**, contact support@pubnub.com.
-
-## License
-
-This project is licensed under the [MIT license].
-
-[MIT license]: https://github.com/pubnub/LICENSE/blob/master/LICENSE
-
