--- crates/pubnub_core/Cargo.toml
@@ -1,19 +1,0 @@
-[package]
-name = "pubnub_core"
-version = "0.0.0"
-edition = "2021"
-license = "MIT"
-authors = ["PubNub <support@pubnub.com>"]
-description = "PubNub SDK for Rust"
-repository = "https://github.com/pubnub/rust"
-documentation = "https://docs.rs/pubnub" # TODO: add docs and connect with pubnub docs
-homepage = "https://www.pubnub.com"
-categories = ["api-bindings", "asynchronous", "network-programming"]
-
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-async-trait = "0.1.66"
-thiserror = "1.0"
-
