--- crates/pubnub_transport_reqwest/Cargo.toml
@@ -1,25 +1,0 @@
-[package]
-name = "pubnub_transport_reqwest"
-version = "0.1.0"
-edition = "2021"
-license = "MIT"
-authors = ["PubNub <support@pubnub.com>"]
-description = "PubNub SDK for Rust"
-repository = "https://github.com/pubnub/rust"
-documentation = "https://docs.rs/pubnub" # TODO: add docs and connect with pubnub docs
-homepage = "https://www.pubnub.com"
-categories = ["api-bindings", "asynchronous", "network-programming"]
-
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-async-trait = "0.1.66"
-pubnub_core = { path = "../pubnub_core" }
-reqwest = "0.11.14"
-log = "0.4.17"
-
-[dev-dependencies]
-tokio = { version = "1", features = ["rt", "macros"]}
-wiremock = "0.5"
-env_logger = "0.10.0"
