+++ src/core/mod.rs
@@ -1,0 +1,13 @@
+//! TODO: docs
+
+pub use error::PubNubError;
+pub mod error;
+
+pub use transport::Transport;
+pub mod transport;
+
+pub use transport_request::{TransportMethod, TransportRequest};
+pub mod transport_request;
+
+pub use transport_response::TransportResponse;
+pub mod transport_response;
